requirements.txt
youtube-transcript-api==0.6.1
openai==1.7.2
python-dotenv==1.0.0
pytube==15.0.0 